Use 32 bit instead of 16

+++ header/MessageReceiver.h
@@ -49,7 +49,7 @@ public:
     void handleUdpData(std::string &&data);
 
 private:
-    bool isAfterCurrentSequenceNum(uint16_t number);
+    bool isAfterCurrentSequenceNum(uint32_t number);
    
 };
